Split off separate header for libev function prototypes
This patch creates a new header - ev-proto.h - which contains all
of the protoypes for libev functions. This allows us to create a
shared object of libuv without exposing libev internal functions.
